Fix review suggestion. Add bir.ksy entry

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/BIRPackageSymbolEnter.java
@@ -416,7 +416,7 @@ public class BIRPackageSymbolEnter {
     private void defineGlobalVarDependencies(BInvokableSymbol invokableSymbol, DataInputStream dataInStream)
             throws IOException {
 
-        long length = dataInStream.readLong();
+        long length = dataInStream.readInt();
         for (int i = 0; i < length; i++) {
             String globalVarName = getStringCPEntryValue(dataInStream.readInt());
             invokableSymbol.dependentGlobalVars.add(this.globalVarMap.get(globalVarName));
